Test store and action for setting token and retrieving shoppinglists
