+++ src/actions/userActions.js
@@ -48,5 +48,4 @@ export function createUser(user) {
 
 export function logOut() {
     localStorage.clear();
-    redirect();
 }
