+++ src/components/helperComponents/LogoutButton.js
@@ -2,6 +2,7 @@ import React, {PropTypes} from 'react';
 import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
 import {logOut} from '../../actions/userActions';
+import {redirect} from "../../helper";
 
 export class LogoutButton extends React.Component{
 
@@ -13,7 +14,9 @@ export class LogoutButton extends React.Component{
     }
 
     logOutUser = (event) => {
-        this.props.logOutUser();
+        this.props.logOutUser().then(() => {
+            redirect();
+        });
     };
 
     render(){
